backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.19
python-dotenv==1.0.1
google-genai==1.0.0
google-search-results==2.4.2
Pillow==11.1.0
